backport of commit 4f14f7bfec975fa7caf93c0b3dd716d12b1913cc
